requirements.txt
Flask==3.0.3
azure-storage-blob==12.22.0
pandas==2.2.2
